refactor(): 클래스이동 및 first-command 분리
