+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,6 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.Color;
+import chess.domain.Score;
 import chess.domain.position.Diagonal;
 import chess.domain.Name;
 import chess.domain.position.Position;
