+++ src/main/java/chess/domain/piece/King.java
@@ -1,6 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.Color;
+import chess.domain.Score;
 import chess.domain.position.Cross;
 import chess.domain.position.Diagonal;
 import chess.domain.Name;
@@ -9,7 +10,7 @@ import chess.domain.position.Position;
 import java.util.ArrayList;
 import java.util.List;
 
-import static chess.domain.piece.Score.ZERO;
+import static chess.domain.Score.ZERO;
 
 public class King extends Piece {
     private static final Position INITIAL_BLACK_POSITION = Position.of('e', '8');
