+++ src/main/java/chess/domain/piece/Piece.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.Color;
 import chess.domain.Name;
+import chess.domain.Score;
 import chess.domain.position.Position;
 
 public abstract class Piece {
